% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as_litecharts4r}
\alias{as_litecharts4r}
\title{As Litecharts}
\usage{
as_litecharts4r(e, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{e}{The {echarts4r} plot object to render.}

\item{width, height}{Dimensions of plot.}

\item{elementId}{ID of htmlwidgets (shadow DOM)}
}
\description{
Render a plot object from {echarts4r} use litecharts4r.
}
