% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{lite_options}
\alias{lite_options}
\alias{set_title}
\alias{set_tooltip}
\alias{set_x_axis}
\alias{set_y_axis}
\alias{add_serie}
\alias{set_legend}
\alias{set_grid}
\alias{set_polar}
\alias{set_radius_axis}
\alias{set_angle_axis}
\alias{add_data_zoom}
\alias{add_visual_map}
\alias{set_axis_pointer}
\alias{set_toolbox}
\alias{set_brush}
\alias{set_geo}
\alias{set_parallel}
\alias{set_parallel_axis}
\alias{set_single_axis}
\alias{set_timeline}
\alias{set_graphic}
\alias{set_calendar}
\alias{set_dataset}
\alias{set_text_style}
\alias{set_state_animations}
\title{Lite Options}
\usage{
set_title(l, ...)

set_tooltip(l, ...)

set_x_axis(l, ...)

set_y_axis(l, ...)

add_serie(l, ...)

set_legend(l, ...)

set_grid(l, ...)

set_polar(l, ...)

set_radius_axis(l, ...)

set_angle_axis(l, ...)

add_data_zoom(l, ...)

add_visual_map(l, ...)

set_axis_pointer(l, ...)

set_axis_pointer(l, ...)

set_toolbox(l, ...)

set_brush(l, ...)

set_geo(l, ...)

set_parallel(l, ...)

set_parallel_axis(l, ...)

set_single_axis(l, ...)

set_timeline(l, ...)

set_graphic(l, ...)

set_calendar(l, ...)

set_dataset(l, ...)

set_text_style(l, ...)

set_state_animations(l, ...)
}
\arguments{
\item{l}{litecharts4r object.}

\item{...}{Options.}
}
\description{
Lite wrapper to build the echarts options.
}
