% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litecharts4r.R
\name{litecharts4r-shiny}
\alias{litecharts4r-shiny}
\alias{litecharts4rOutput}
\alias{renderLitecharts4r}
\title{Shiny bindings for litecharts4r}
\usage{
litecharts4rOutput(outputId, width = "100\%", height = "400px")

renderLitecharts4r(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a litecharts4r}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using litecharts4r within Shiny
applications and interactive Rmd documents.
}
