% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litecharts4r.R
\name{litecharts4r}
\alias{litecharts4r}
\title{Chrat}
\usage{
litecharts4r(..., width = "100\%", height = NULL, elementId = NULL)
}
\arguments{
\item{...}{Any first-level options.}

\item{width, height}{Dimensions of plot.}

\item{elementId}{ID of htmlwidgets (shadow DOM)}
}
\description{
Initialise a chart.
}
